+++ src/main/java/com/example/logistic/model/dtos/AdminDTO.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class AdminDTO extends PersonaDTO{
-    public AdminDTO(Integer id, String name, String lastName, String username, Integer tenantId) {
+    public AdminDTO(Long id, String name, String lastName, String username, Long tenantId) {
         super(id, name, lastName, username, tenantId);
     }
 }
