+++ src/main/java/com/example/logistic/model/dtos/VendedorDTO.java
@@ -8,7 +8,7 @@ import java.util.List;
 @Getter
 @Setter
 public class VendedorDTO extends PersonaDTO{
-    public VendedorDTO(Integer id, String nombre, String apellido, String username, Integer tenantId) {
+    public VendedorDTO(Long id, String nombre, String apellido, String username, Long tenantId) {
         super(id, nombre, apellido, username, tenantId);
     }
