+++ src/main/java/com/example/logistic/repository/RutaRepository.java
@@ -10,12 +10,12 @@ import java.util.Date;
 import java.util.List;
 
 @Repository
-public interface RutaRepository extends JpaRepository<Ruta, Integer> {
+public interface RutaRepository extends JpaRepository<Ruta, Long> {
     // Consulta para obtener la ruta por el id del driver
     // Consulta para obtener la ruta del driver por fecha actual
     @Query("SELECT r FROM Ruta r WHERE r.driver.id = :driverId AND r.date = :date AND r.completada = false")
-    Ruta findByDriverId(@Param("driverId") Integer driverId, @Param("date") Date date);
+    Ruta findByDriverId(@Param("driverId") Long driverId, @Param("date") Date date);
 
     @Query("SELECT r FROM Ruta r WHERE r.date = :date and r.tenant.id = :tenantId")
-    List<Ruta> findByDay(@Param("date") Date date ,@Param("tenantId") Integer tenantId);
+    List<Ruta> findByDay(@Param("date") Date date ,@Param("tenantId") Long tenantId);
 }
